[refactor] #128  회의 진행 후 네이밍 관련 수정! -> facade 붙이기보단 전략의 맥락이더 중요하여 변경
